+++ packages/frontend/src/App.routes.tsx
@@ -3,7 +3,7 @@ export const CREATE_STREAM = 'create-stream';
 export const ACCOUNT = 'account';
 export const STREAM = 'stream/:id';
 
-export const menu = {
+export const routes = {
   Streams: {
     url: `${STREAMS}`,
   },
