+++ packages/frontend/src/components/index.ts
@@ -1,3 +1,4 @@
 export * from './Header';
 export * from './Footer';
 export * from './Modal';
+export * from './Loader';
