--- packages/frontend/src/features/StreamTeasers/atoms.ts
@@ -1,3 +1,0 @@
-import { atom } from 'jotai';
-
-export const ATOM_STREAM_TEASERS_LIST = atom<any>(null);
