+++ packages/frontend/src/features/StreamTeasers/components/StreamTeasersList/StreamTeasersList.module.scss
@@ -20,14 +20,6 @@
     margin-bottom: 40px;
 }
 
-.title {
-
-}
-
-.search {
-
-}
-
 .content {
     width: 100%;
     display: flex;
