+++ packages/frontend/src/features/Wallet/components/WalletModal/WalletModal.interface.ts
@@ -1,0 +1,3 @@
+export interface WalletModalProps {
+  onClose: () => void;
+}
