+++ packages/frontend/src/features/Wallet/hooks.ts
@@ -1,0 +1,58 @@
+import { useEffect, useState } from 'react';
+import { getProgramMetadata, ProgramMetadata } from '@gear-js/api';
+import { useAlert, useAccount } from '@gear-js/react-hooks';
+import { Buffer } from 'buffer';
+import { LOCAL_STORAGE } from '@/consts';
+import { WALLET } from './consts';
+import { WalletId } from './types';
+
+function useMetadata(source: RequestInfo | URL) {
+  const [data, setData] = useState<ProgramMetadata>();
+
+  useEffect(() => {
+    fetch(source)
+      .then((res) => res.text() as Promise<string>)
+      .then((raw) => getProgramMetadata(`0x${raw}`))
+      .then((meta) => setData(meta));
+  }, [source]);
+
+  return data;
+}
+
+function useWasmMetadata(source: RequestInfo | URL) {
+  const alert = useAlert();
+  const [data, setData] = useState<Buffer>();
+
+  useEffect(() => {
+    if (source) {
+      fetch(source)
+        .then((response) => response.arrayBuffer())
+        .then((array) => Buffer.from(array))
+        .then((buffer) => setData(buffer))
+        .catch(({ message }: Error) => alert.error(`Fetch error: ${message}`));
+    }
+  }, [source, alert]);
+
+  return { buffer: data };
+}
+
+function useWallet() {
+  const { accounts } = useAccount();
+
+  const [walletId, setWalletId] = useState<WalletId | undefined>(localStorage[LOCAL_STORAGE.WALLET]);
+
+  const resetWalletId = () => setWalletId(undefined);
+
+  const getWalletAccounts = (id: WalletId) => accounts.filter(({ meta }) => meta.source === id);
+
+  const saveWallet = () => walletId && localStorage.setItem(LOCAL_STORAGE.WALLET, walletId);
+
+  const removeWallet = () => localStorage.removeItem(LOCAL_STORAGE.WALLET);
+
+  const wallet = walletId && WALLET[walletId];
+  const walletAccounts = walletId && getWalletAccounts(walletId);
+
+  return { wallet, walletAccounts, setWalletId, resetWalletId, getWalletAccounts, saveWallet, removeWallet };
+}
+
+export { useMetadata, useWasmMetadata, useWallet };
