--- packages/frontend/src/features/Wallet/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from './useWallet.hook';
-export * from './useMetadata.hook';
