+++ packages/frontend/src/features/Wallet/types.ts
@@ -1,5 +1,10 @@
 import { WALLET } from './consts';
 
-type WalletId = keyof typeof WALLET;
+export type WalletId = keyof typeof WALLET;
 
-export type { WalletId };
+export type WalletValue = {
+  name: string;
+  SVG: string;
+};
+
+export type WalletEntry = [WalletId, WalletValue];
