+++ packages/frontend/src/ui/Pagination/Pagination.interface.ts
@@ -1,0 +1,6 @@
+export interface PaginationProps {
+  totalRows: number;
+  rowsPerPage: number;
+  currentPage: number;
+  setCurrentPage: () => void;
+}
