+++ packages/frontend/src/ui/Pagination/Pagination.module.scss
@@ -1,0 +1,11 @@
+.pagination {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 0 20px;
+
+    &-right {
+        display: flex;
+        gap: 24px;
+    }
+}
