+++ packages/frontend/src/ui/Table/Table.module.scss
@@ -69,18 +69,6 @@
     align-items: center;
 }
 
-.pagination {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 0 20px;
-
-    &-right {
-        display: flex;
-        gap: 24px;
-    }
-}
-
 .empty-table {
     display: flex;
     flex-direction: column;
