+++ packages/frontend/src/ui/Table/Table.tsx
@@ -1,10 +1,5 @@
 import { useEffect, useState } from 'react';
-import { Button } from '@ui';
 import { cx } from '@/utils';
-import vectorLeft from '@/assets/icons/vector-left.svg';
-import vectorRight from '@/assets/icons/vector-right.svg';
-import doubleVectorLeft from '@/assets/icons/double-vector-left.svg';
-import doubleVectorRight from '@/assets/icons/double-vector-right.svg';
 import styles from './Table.module.scss';
 import {
   TableBodyProps,
@@ -15,63 +10,7 @@ import {
   TableRow,
   TableRowProps,
 } from './Table.interfaces';
-
-function Pagination({ totalRows, rowsPerPage, currentPage, setCurrentPage }: any) {
-  const [pageCount, setPageCount] = useState(Math.ceil(totalRows / rowsPerPage));
-
-  useEffect(() => {
-    const pages = Math.ceil(totalRows / rowsPerPage);
-
-    if (pages && currentPage > pages) {
-      setCurrentPage(pages);
-    }
-
-    if (!pages) {
-      setCurrentPage(1);
-    }
-
-    setPageCount(pages);
-  }, [totalRows, rowsPerPage, currentPage, setCurrentPage]);
-
-  return (
-    <div className={cx(styles.pagination)}>
-      <div className={cx(styles['pagination-left'])}>out of {totalRows}</div>
-      <div className={cx(styles['pagination-right'])}>
-        <Button
-          variant="icon"
-          icon={doubleVectorLeft}
-          label=""
-          disabled={currentPage < 2}
-          onClick={() => setCurrentPage(1)}
-        />
-        <Button
-          variant="icon"
-          icon={vectorLeft}
-          label=""
-          disabled={currentPage < 2}
-          onClick={() => setCurrentPage(currentPage - 1)}
-        />
-        <span>{currentPage}</span>
-        <span>out of</span>
-        <span>{pageCount || 1}</span>
-        <Button
-          variant="icon"
-          icon={vectorRight}
-          label=""
-          disabled={currentPage > pageCount - 1}
-          onClick={() => setCurrentPage(currentPage + 1)}
-        />
-        <Button
-          variant="icon"
-          icon={doubleVectorRight}
-          label=""
-          disabled={currentPage > pageCount - 1}
-          onClick={() => setCurrentPage(pageCount)}
-        />
-      </div>
-    </div>
-  );
-}
+import { Pagination } from '../Pagination';
 
 function Cell({ className, children }: TableCellProps) {
   return <th className={cx(styles.cell, className)}>{children}</th>;
@@ -140,10 +79,12 @@ function Table({ rows, columns, pagination, renderCell, renderHeaderCell, classN
                 ))}
               </>
             ) : (
-              <div className={cx(styles['empty-table'])}>
-                <h3 className={cx(styles['empty-table-title'])}>No subscribers</h3>
-                <span className={cx(styles['empty-table-caption'])}>You don&apos;t have subscribers yet ...</span>
-              </div>
+              <tr className={cx(styles['empty-table'])}>
+                <td>
+                  <h3 className={cx(styles['empty-table-title'])}>No subscribers</h3>
+                  <span className={cx(styles['empty-table-caption'])}>You don&apos;t have subscribers yet ...</span>
+                </td>
+              </tr>
             )}
           </Body>
         </table>
