--- packages/frontend/src/utils/cx.ts
@@ -1,3 +1,0 @@
-import clsx from 'clsx';
-
-export const cx = (...styles: string[]) => clsx(...styles);
