--- packages/frontend/src/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './cx';
-export * from './copyToClipboard';
