--- packages/frontend/src/utils/socket.js
@@ -1,3 +1,0 @@
-import { io } from 'socket.io-client';
-
-export default io('http://127.0.0.1:3001');
